services/detection/requirements.txt
# ══════════════════════════════════════════════════════════════
# DETECTION SERVICE - Dependencies
# ══════════════════════════════════════════════════════════════

# --------------------------------------------------------------
# YOLO / Ultralytics - Object Detection
# --------------------------------------------------------------
ultralytics>=8.0.0              # YOLO model framework

# --------------------------------------------------------------
# OpenCV - Image Processing
# --------------------------------------------------------------
opencv-python-headless>=4.8.0   # Image manipulation, drawing boxes

# --------------------------------------------------------------
# RabbitMQ Client
# --------------------------------------------------------------
pika>=1.3.0                     # Message queue communication

# --------------------------------------------------------------
# PostgreSQL Client
# --------------------------------------------------------------
psycopg2-binary>=2.9.0          # PostgreSQL database driver
                                # "binary" = pre-compiled, no build tools needed

# --------------------------------------------------------------
# Utilities
# --------------------------------------------------------------
python-dotenv>=1.0.0            # Environment variable loading
numpy>=1.24.0                   # Array operations (used by OpenCV/YOLO)
lap>=0.5.0                      # Linear Assignment Problem (for tracking)
